+++ include/mn_auto_version.hpp
@@ -3,21 +3,27 @@
 
 namespace mn{
 	
+	//Date Version Types
+	#define MN_VERSION_DATE "23"
+	#define MN_VERSION_MONTH "05"
+	#define MN_VERSION_YEAR "2021"
+	#define MN_VERSION_UBUNTU_VERSION_STYLE  "21.05"
+	
 	//Software Status
-	#define MN_VERSION_STATUS  ""
-	#define MN_VERSION_STATUS_SHORT  ""
+	#define MN_VERSION_STATUS  "Beta"
+	#define MN_VERSION_STATUS_SHORT  "beta"
 	
 	//Standard Version Type
-	#define MN_VERSION_MAJOR  0
-	#define MN_VERSION_MINOR  0
-	#define MN_VERSION_BUILD  3
-	#define MN_VERSION_REVISION  1113
+	#define MN_VERSION_MAJOR  2
+	#define MN_VERSION_MINOR  295
+	#define MN_VERSION_BUILD  8908
+	#define MN_VERSION_REVISION  8908
 	
 	//Miscellaneous Version Types
-	#define MN_VERSION_BUILDS_COUNT  4
-	#define MN_VERSION_RC_FILEVERSION 0,0,3,1113
-	#define MN_VERSION_RC_FILEVERSION_STRING "0, 0, 3, 1113\0"
-	#define MN_VERSION_FULLVERSION_STRING  "0.0.3.1113"
+	#define MN_VERSION_BUILDS_COUNT  206
+	#define MN_VERSION_RC_FILEVERSION 2,295,8908,8908
+	#define MN_VERSION_RC_FILEVERSION_STRING "2, 295, 8908, 8908\0"
+	#define MN_VERSION_FULLVERSION_STRING  "2.295.8908.8908"
 	
 	//These values are to keep track of your versioning state, don't modify them.
 	#define MN_VERSION_BUILD_HISTORY  0