+++ include/mn_functional.hpp
@@ -28,157 +28,6 @@
 using nullptr_t = decltype(nullptr);
 
 namespace mn {
-	template <typename T>
-	struct type_traits {
-		using value_type = T;
-		using const_type = const T;
-		using reference = T&;
-		using const_reference = const T&;
-		using pointer = T*;
-		using const_pointer = const T*;
-	};
-
-	template <typename T>
-	struct type_traits<T&> {
-		using value_type = T;
-		using const_type = const T;
-		using reference = T&;
-		using const_reference = const T&;
-		using pointer = T*;
-		using const_pointer = const T*;
-	};
-
-	template <typename T>
-	struct type_traits<const T> {
-		using value_type = T;
-		using const_type = const T;
-		using reference = T&;
-		using const_reference = const T&;
-		using pointer = T*;
-		using const_pointer = const T*;
-	};
-
-	template <typename T>
-	struct type_traits<const T&> {
-		using value_type = T;
-		using const_type = const T;
-		using reference = T&;
-		using const_reference = const T&;
-		using pointer = T*;
-		using const_pointer = const T*;
-	};
-#if 0
-	template<typename Arg, typename TResult>
-	struct unary_function {
-	  /// @c argument_type is the type of the argument
-	  using argument_type = Arg;
-	  /// @c result_type is the return type
-	  using result_type =  TResult;
-	};
-
-
-	template<typename Arg1, typename Arg2, typename TResult>
-	struct binary_function {
-		/// @c first_argument_type is the type of the first argument
-		using first_argument_type = Arg;
-		/// @c second_argument_type is the type of the second argument
-		using second_argument_type = Arg2;
-		/// @c result_type is the return type
-		using result_type =  TResult;
-	};
-
-	template<typename T>
-	struct negate : public unary_function<T, T> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a) const noexcept { return -a; }
-	};
-
-	template<typename T>
-	struct plus : public binary_function<T, T, T> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const  noexcept { return a + b; }
-	};
-
-	template<typename T>
-	struct minus : public binary_function<T, T, T> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a - b; }
-	};
-
-	template<typename T>
-	struct multiplies : public binary_function<T, T, T> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a * b; }
-	};
-
-	template<typename T>
-	struct divides : public binary_function<T, T, T> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a / b; }
-	};
-
-	template<typename T>
-	struct modulus : public binary_function<T, T, T> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a % b; }
-	};
-
-	template<typename T>
-	struct equal_to : public binary_function<T, T, bool> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a == b; }
-	};
-
-	template<typename T>
-	struct not_equal_to : public binary_function<T, T, bool> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a != b; }
-	};
-
-	template<typename T>
-	struct greater : public binary_function<T, T, bool> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a > b; }
-	};
-
-	template<typename T>
-	struct less : public binary_function<T, T, bool> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a < b; }
-	};
-
-	template<typename T>
-	struct greater_equal : public binary_function<T, T, bool> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a >= b; }
-	};
-
-	template<typename T>
-	struct less_equal : public binary_function<T, T, bool> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a <= b; }
-	};
-
-	template<typename T>
-	struct logic_and : public binary_function<T, T, bool> {
-		typename binary_function<T, T, T>::result_type
-		bool operator()(const T& a, const T& b) const noexcept { return a && b; }
-	};
-
-	template<typename T>
-	struct logic_or : public binary_function<T, T, bool> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a, const T& b) const noexcept { return a || b; }
-	};
-
-	template<typename T>
-	struct logic_not : public unary_function<T, bool> {
-		typename binary_function<T, T, T>::result_type
-		operator()(const T& a) const noexcept { return !a; }
-	};
-#endif // 0
-
-
     template <typename T> struct tag { using type = T; };
     template <typename Tag> using type_t = typename Tag::type;
 
@@ -203,23 +52,6 @@ namespace mn {
 	struct bit_or<A, B, C, N...>
 		: public conditional<A::value, B, bit_or<B, C, N...>>::type { };
 
-    template<typename...>
-    struct bit_and;
-
-	template<>
-	struct bit_and<> : public true_type { };
-
-	template<typename B>
-	struct bit_and<B> : public B { };
-
-	template<typename A, typename B>
-	struct bit_and<A, B> : public conditional<A::value, B, A>::type { };
-
-	template<typename A, typename B, typename C, typename... N>
-	struct bit_and<A, B, C, N...> : public conditional<A::value, bit_and<B, C, N...>, A>::type { };
-
-	template<typename A>
-	struct bit_not : public integral_constant<bool, !A::value> { };
 
     template <typename T> T &&move(T &t) {
         return static_cast<T &&>(t); }
@@ -355,9 +187,6 @@ namespace mn {
     template <class T>
     struct enable_if<true, T> : tag<T> {};
 
-    template<typename... _Cond>
-    using require = typename enable_if<bit_and<_Cond...>::value>::type;
-
     template <bool b, class T = void>
     using enable_if_t = type_t<enable_if<b, T>>;
 
@@ -380,6 +209,20 @@ namespace mn {
     struct is_base_of : public integral_constant<bool, __is_base_of(TBase, TDerived)> { };
 
 
+    template<typename T>
+	struct logic_and {
+		bool operator()(const T& a, const T& b) const noexcept { return a && b; }
+	};
+
+	template<typename T>
+	struct logic_or {
+		bool operator()(const T& a, const T& b) const noexcept { return a || b; }
+	};
+
+	template<typename T>
+	struct logic_not {
+		bool operator()(const T& a) const noexcept { return !a; }
+	};
 }
 
 #endif